backend/src/tests/requirements.txt
auto_mix_prep==0.2.0
colorama==0.4.6
httpx==0.28.1
python-dotenv==1.1.0
tabulate==0.9.0
tools==1.0
